requirements.txt
anthropic>=0.40.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
web3>=6.15.0
tweepy>=4.14.0
python-dotenv>=1.0.0
rich>=13.7.0
requests>=2.31.0
python-telegram-bot>=20.0
slowapi>=0.1.9
